// web/src/gen/proto/termbridge/common/v1/common.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.8
//   protoc               unknown
// source: termbridge/common/v1/common.proto

/* eslint-disable */

export const protobufPackage = 'termbridge.common.v1'

export interface ErrorResp {
  code: string
  error: string
  request_id: string
  details: { [key: string]: any } | undefined
}

export interface HealthResp {
  status: string
}